+++ Avatars/Runtime/Scripts/Configuration/IAvatarConfig.cs
@@ -4,6 +4,9 @@ using UnityEngine;
 
 namespace SPACS.SDK.Avatars
 {
+    /// <summary>
+    /// Each avatar configuration object must implement this class
+    /// </summary>
     public interface IAvatarConfig
     {