+++ Avatars/Runtime/Scripts/Controllers/IAvatarController.cs
@@ -4,13 +4,35 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-using UnityEngine;
-
 namespace SPACS.SDK.Avatars
 {
+    /// <summary>
+    /// Avatar controller interface
+    /// </summary>
     public interface IAvatarController
     {
-        Task Setup(CharacterControllerBase characterController);
+        #region Properties
+
+        public CharacterControllerBase SourceController { get; }
+
+        #endregion
+
+        #region Public API
+
+        /// <summary>
+        /// Setups the avatar controller base given a source controller. 
+        /// The source controller is the object which the avatar is attached to.
+        /// </summary>
+        /// <param name="source">The source controller wich the htis avatar controller is attached to.</param>
+        /// <returns>Task</returns>
+        Task Setup(CharacterControllerBase source);
+
+        /// <summary>
+        /// Unsetups the current character controller.
+        /// </summary>
+        /// <returns>Task</returns>
         Task Unsetup();
+
+        #endregion
     }
 }