+++ CharacterController/Runtime/CharacterControllerSystem.cs
@@ -23,12 +23,18 @@ namespace SPACS.SDK.CharacterController
 
         #region Properties
 
+        /// <summary>
+        /// Characte controller instance
+        /// </summary>
         public CharacterControllerBase CharacterControllerInstance { get; protected set; }
 
         #endregion
 
         #region Unity Events
 
+        /// <summary>
+        /// Invoked when the character controller setup is completed
+        /// </summary>
         public UnityEvent<CharacterBase> OnCharacterControllerSetupComplete { get; } = new();
 
         #endregion
@@ -38,35 +44,42 @@ namespace SPACS.SDK.CharacterController
         public override void Init()
         {
             if (spawnCharacterOnInit)
-                Spawn();
+                CreateCharacterControllerInstance();
         }
 
         #endregion
 
         #region Public API
 
-        public virtual void Spawn() {
-
-            if (!characterControllerPrefab) {
-                CharacterControllerInstance = FindObjectOfType<CharacterControllerBase>();
-            } else {
-                if (!Instantiate(characterControllerPrefab, spawnPose.position, spawnPose.rotation).TryGetComponent(out CharacterControllerBase characterController)) {
-                    throw new Exception("Character controller not found");
-                }
-                CharacterControllerInstance = characterController;
-            }
-
+        /// <summary>
+        /// Creates the character controller instance. 
+        /// A character controller can be referenced as a prefab through the serialized field `characterControllerPrefab`, or can be already in scene
+        /// </summary>
+        public virtual void CreateCharacterControllerInstance() 
+        {
+            CharacterControllerInstance = characterControllerPrefab 
+                ? Instantiate(characterControllerPrefab, spawnPose.position, spawnPose.rotation).GetComponent<CharacterControllerBase>()
+                : CharacterControllerInstance = FindObjectOfType<CharacterControllerBase>();
         }
 
+        /// <summary>
+        /// Destroys the character controller instance
+        /// </summary>
+        public virtual void DestroyCharacterControllerInstance()
+        {
+            if (CharacterControllerInstance != null)
+                Destroy(CharacterControllerInstance.gameObject);
+        }
+
+        /// <summary>
+        /// Moves the character controller to a destination position and rotation
+        /// </summary>
+        /// <param name="newPose">The destination pose of the character</param>
         public virtual void MoveCharacter(Pose newPose)
         {
             CharacterControllerInstance.transform.SetPositionAndRotation(newPose.position, newPose.rotation);
         }
 
-        public virtual void Destroy() {
-            if (CharacterControllerInstance != null)
-                Destroy(CharacterControllerInstance.gameObject);
-        }
 
         #endregion
     }