+++ Transitions/Runtime/Scripts/NoTransitionProvider.cs
@@ -1,0 +1,25 @@
+﻿using System.Threading.Tasks;
+using UnityEngine;
+
+namespace SPACS.SDK.Transitions
+{
+    public class NoTransitionProvider : AbstractTransitionProvider
+    {
+        [SerializeField, Tooltip("The GameObject to activate")]
+        private GameObject content;
+
+        public override async Task EnterTransitionAsync()
+        {
+            await Task.Yield();
+            content.SetActive(true);
+            await Task.Yield();
+        }
+
+        public override async Task ExitTransitionAsync()
+        {
+            await Task.Yield();
+            content.SetActive(false);
+            await Task.Yield();
+        }
+    }
+}